# openapi.yaml
openapi: 3.0.0
info:
  title: 'Mp3 api'
  version: '0.1'
paths:
...
  /tag/attach-tag:
    post:
      tags:
        - Tag
      summary: 'Прикрепление тэга к посту'
      description: 'Прикрепление тэга к посту'
      operationId: f878841572804e03bf1bca100c619658
      requestBody:
        description: 'Данные для прикрепление'
        required: true
        content:
          application/json:
            schema:
              required:
                - id_post
                - id_tag
              properties:
                id_post:
                  type: integer
                  example: 1
                id_tag:
                  type: integer
                  example: 2
              type: object
      responses:
        '200':
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                type: object
        '400':
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  errors: { type: object, example: { tag_id: 'Tag ID не найден' } }
                type: object
  /tag/detach-tag:
    delete:
      tags:
        - Tag
      summary: 'Открепление тэга от поста'
      description: 'Открепление тэга от поста'
      operationId: e095760cccc9f1eedab70327ead66641
      requestBody:
        description: 'Данные для открепления'
        required: true
        content:
          application/json:
            schema:
              required:
                - id_post
                - id_tag
              properties:
                id_post:
                  type: integer
                  example: 1
                id_tag:
                  type: integer
                  example: 2
              type: object
      responses:
        '200':
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                type: object
        '400':
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  errors: { type: object, example: { tag_id: 'Tag ID не найден' } }
                type: object
  /tag:
    get:
      tags:
        - Tag
      summary: 'Получить список всех тэгов'
      description: 'Возвращает список всех тэгов с возможностью фильтрации'
      operationId: 91fb7d06e13fa31094e542c0f866a705
      parameters:
        -
          name: page
          in: query
          description: 'Номер страницы'
          required: false
          schema:
            type: integer
            example: 1
        -
          name: pageSize
          in: query
          description: 'Количество элементов на страницы'
          required: false
          schema:
            type: integer
            example: 5
        -
          name: sort
          in: query
          description: Сортировка
          required: false
          schema:
            type: string
            example: id
      responses:
        '200':
          description: 'Успешный ответ'
          content:
            application/json:
              schema:
                properties:
                  id: { type: integer, example: 1 }
                  title: { type: string, example: testr }
                type: object
    post:
      tags:
        - Tag
      summary: 'Создаёт новый тэг'
      description: 'Создание нового тэга '
      operationId: b8bf26b2c8c74e5a43db501ebd23843b
      requestBody:
        description: 'Данные для создания'
        required: true
        content:
          application/json:
            schema:
              required:
                - title
              properties:
                title:
                  type: string
                  example: Test
              type: object
      responses:
        '200':
          description: 'Успешное удаление'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  id: { type: integer, example: 1 }
                  title: { type: string, example: testTag }
                type: object
        '400':
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  errors: { type: object, example: { title: 'Не может быть пустым' } }
                type: object
  '/tag/{id}':
    put:
      tags:
        - Tag
      summary: 'Изменяет  тэг'
      description: 'Изменяет  тэг '
      operationId: fdeff039de4d66c5e032752a74ad9616
      parameters:
        -
          name: id
          in: query
          description: 'Id тэга'
          required: false
          schema:
            type: integer
            example: 1
      requestBody:
        description: 'Данные для редактирования'
        required: true
        content:
          application/json:
            schema:
              required:
                - title
              properties:
                title:
                  type: string
                  example: Test
              type: object
      responses:
        '200':
          description: 'Успешное изменение'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  id: { type: integer, example: 1 }
                  message: { type: string, example: 'Тэг успешно изменён' }
                type: object
        '400':
          description: 'Ошибка валидации'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: false }
                  errors: { type: object, example: { title: 'Не может быть пустым' } }
                type: object
    delete:
      tags:
        - Tag
      summary: 'Удаляет  тэг'
      description: 'Удаляет  тэг '
      operationId: 7b2a5c6b87f7294b2a7fe783b5406524
      parameters:
        -
          name: id
          in: query
          description: 'Id тэга'
          required: false
          schema:
            type: integer
            example: 1
      responses:
        '200':
          description: 'Успешное удаление'
          content:
            application/json:
              schema:
                properties:
                  success: { type: boolean, example: true }
                  id: { type: integer, example: 1 }
                  message: { type: string, example: 'Тэг успешно удалён' }
                type: object
...